% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_logit.R
\name{logit}
\alias{logit}
\alias{logit.numeric}
\alias{logit.formula}
\alias{logit.default}
\title{Logistic Regression}
\usage{
logit(...)

\method{logit}{numeric}(X, y, ...)

\method{logit}{formula}(formula, data, na.action = na.fail, ...)

\method{logit}{default}(others, ...)
}
\arguments{
\item{...}{arguments differ for each method. Either a formula with a data frame
or a numeric design matrix and response vector.}

\item{X}{a design matrix of dimension n * p}

\item{y}{a vector of observation of length n}

\item{formula}{an object of class "formula" (or one that can be coerced
to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object
coercible by as.data.frame to a data frame) containing the variables in the
model. If not found in data, the variables are taken from environment(formula),
typically the environment from which lm is called.}

\item{na.action}{a function to handle NAs in the data. The default is set as
na.fail}

\item{others}{objects that is neither of class formula nor of numeric matrix}
}
\value{
\code{logit} returns an object of class "logit", a list with
the following componets:

\item{coefficients}{a named vector of coefficients}
\item{linear.predictors}{linear fit}
\item{fitted.values}{the fitted mean values or fitted conditional probability.}
\item{residuals}{the working residuals.}
\item{AIC}{A version of Akaike's An Information Criterion, minus twice
the maximized log-likelihood plus twice the number of parameters.}
\item{call}{the matched call}
\item{df.null}{the residual degrees of freedom for the null model, i.e. n - 1}
\item{df.residual}{the residual degrees of freedom, i.e. n - p}
\item{null.deviance}{The deviance for the null model, minus two times the
log-likelihood of the null model. This is to compare with \code{deviance}.}
\item{deviance}{minus two times the log-likelihood of the saturated model}
\item{model}{the model frame}
\item{y}{the binary response vector}
\item{X}{the model matrix used for the regression, aliased variables omitted.}
\item{X_ini}{initial model matrix given either by formula or as matrix,
including aliased variables}
\item{formula}{The given formula in case of calling \code{logit} using an
argument of class \code{formula}}
}
\description{
Running a logistic regression given binary dependent variable
}
\details{
\code{logit} can perform with inputs of class \code{formula} and
\code{data.frame} like \code{logit(response ~ terms, data = data.frame)}
as well as with numeric response vector and design matrix like
\code{logit(X, y)}. Terms in the formula can be specified using '+',
':'(interaction terms), and also '*'(cross terms).

When calling \code{logit} with formula, the intercept term will be automatically
added, which can be taken out by adding "- 1 (or any constant)" in the formula
(@seealso model.matrix).

If there exist two terms with exact same entries, one of them is going to be
automatically deleted to avoid multicollinearity which would cause unstable
estimates.
}
\examples{
#' set.seed(1)
y <- rbinom(n = 100, size = 1, prob = 0.5)
X <- cbind(rep(1, length(y)), rnorm(n = 100))
colnames(X) <- c("(Intercept)", "X")
logit(X, y)$coefficients
#'                    [,1]
#' (Intercept) -0.07798872
#' X            0.13041887
#' set.seed(1)
y <- rbinom(n = 100, size = 1, prob = 0.5)
X <- rnorm(n = 100)
data <- data.frame(y = y, X = X)
logit(y ~ X, data)$coefficients
#'                     [,1]
#'  (Intercept) -0.07798872
#'  X            0.13041887
}
\references{
R Core Team (2018). R documentation for stats::glm. R: A language and environment for statistical
computing. R Foundation for Statistical Computing, Vienna, Austria.
URL https://www.R-project.org/.
}
