% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plotMethod.R
\name{plot.logit}
\alias{plot.logit}
\title{Plot method for S3 class \code{logit}}
\usage{
\method{plot}{logit}(x, ...)
}
\arguments{
\item{x}{object of class \code{logit}}

\item{...}{further arguments to be passed to plot}
}
\description{
Plot method for S3 class \code{logit}
}
\details{
For objects of class \code{logit}, plot method returns 4 plots in total, namely
Regression result, Residuals vs. Fitted, Normal Q-Q and Scale-Location plot.
}
\references{
R Core Team (2018). R documentation for \code{stats::plot.lm}. R: A language and environment for statistical
computing. R Foundation for Statistical Computing, Vienna, Austria.
URL https://www.R-project.org/.
}
