% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_summaryMethod.R
\name{summary.logit}
\alias{summary.logit}
\title{Summary Method for S3 class \code{logit}}
\usage{
\method{summary}{logit}(object, ...)
}
\arguments{
\item{object}{object of class \code{logit}}

\item{...}{further arguments to be passed to print}
}
\value{
\code{summary.logit} returns an object of class "summary.logit", a list with
the following componets:

\item{call}{The component from the \code{object}}
\item{deviance}{The component from the \code{object}}
\item{AIC}{The component from the \code{object}}
\item{df.residual}{The component from the \code{object}}
\item{null.deviance}{The component from the \code{object}}
\item{df.null}{The component from the \code{object}}
\item{deviance.resid}{The deviance residuals}
\item{coefficients}{The matrix of coefficients, standard errors, z-values and
 p-values. Aliased coefficients are omitted.}
\item{df}{A vector consisting of the rank of the model and the number of
residual degrees of freedom as well as the number of coefficients (including
aliased ones).}
\item{covariance}{The estimated covariance matrix of the estimated coefficients.}
}
\description{
Summary Method for S3 class \code{logit}
}
\details{
Summary method for class \code{logit} returns not only selected components of
object \code{logit} that are important for regression analysis - e.g. null/residual
deviance, AIC, degrees of freedom null/residuals, but also a few additional
elements like deviance residuals and covariance matrix of the coefficient
estimates. Moreover, the \code{coefficients} component provides a matrix of
coefficient estimates, their standard errors, z-values and p-values.
}
\references{
R Core Team (2018). R documentation for \code{stats::summary.glm}. R: A language and environment for statistical
computing. R Foundation for Statistical Computing, Vienna, Austria.
URL https://www.R-project.org/.
}
